requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
bandit==1.8.6
blinker==1.9.0
boltons==21.0.0
bracex==2.6
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-option-group==0.5.9
colorama==0.4.6
dash==3.2.0
dash-bootstrap-components==2.0.4
dash_ag_grid==32.3.2
dash_treeview_antd==0.0.1
exceptiongroup==1.2.2
face==24.0.0
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.45
glom==22.1.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jschema-to-python==1.2.3
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
linkify-it-py==2.0.3
mando==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.16.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.10.1
nest-asyncio==1.6.0
numpy==2.3.4
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pbr==7.0.2
peewee==3.18.3
platformdirs==4.5.0
plotly==6.3.1
prompt_toolkit==3.0.52
protobuf==6.33.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
radon==6.0.1
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rich==13.5.3
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
sarif-om==1.0.4
semgrep==1.142.0
setuptools==80.9.0
shellingham==1.5.4
shiny==1.5.0
shinychat==0.2.8
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
stevedore==5.5.0
tomli==2.0.2
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.38.0
vulture==2.14
watchfiles==1.1.1
wcmatch==8.5.2
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
